.aws-sam/cache/1bdf58fc-4cb2-4d30-b4b1-5da1c3e6e87c/requirements.txt
aws-msk-iam-sasl-signer-python
confluent_kafka